% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.location.R
\name{as.location}
\alias{as.location}
\title{Convert a path or URL to a location object.}
\usage{
as.location(x, ...)
}
\arguments{
\item{x}{Input.}

\item{...}{Ignored.}
}
\description{
Convert a path or URL to a location object.
}
\examples{
\dontrun{
# A file
file <- system.file("examples", "zillow_or.geojson",
  package = "geojsonlint")
as.location(file)

# A URL
url <- paste0("https://raw.githubusercontent.com/glynnbird/",
  "usstatesgeojson/master/california.geojson")
as.location(url)
}
}
