% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonlint.R
\name{geojson_lint}
\alias{geojson_lint}
\title{Validate GeoJSON using geojsonlint.com web service}
\usage{
geojson_lint(...)
}
\arguments{
\item{...}{ignored}
}
\description{
Validate GeoJSON using geojsonlint.com web service
}
