% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_validate.R
\name{geojson_validate}
\alias{geojson_validate}
\title{geojson_validate}
\usage{
geojson_validate(x, verbose = FALSE, greedy = FALSE, error = FALSE)
}
\arguments{
\item{x}{Input, a geojson character string, json object, or file or
url pointing to one of the former}

\item{verbose}{TRUE or FALSE}

\item{greedy}{TRUE or FALSE}

\item{error}{TRUE or FALSE}
}
\value{
TRUE or FALSE
}
\description{
geojson_validate
}

