% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonlint-package.r
\docType{package}
\name{geojsonlint-package}
\alias{geojsonlint-package}
\alias{geojsonlint}
\title{\strong{GeoJSON Linting}}
\description{
\strong{GeoJSON Linting}
}
\section{Package API}{

\itemize{
\item \code{\link[=geojson_hint]{geojson_hint()}} - Checks validity of geojson using the Javascript library
\code{geojsonhint}
\item \code{\link[=geojson_validate]{geojson_validate()}} - Checks validity of geojson using a GeoJSON schema
and the Javascript library \code{is-my-json-valid}
}
}

\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

Andy Teucher \email{andy.teucher@gmail.com}
}
