% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{gj_write}
\alias{gj_write}
\title{Write inputs to a geojson file}
\usage{
gj_write(x, file, ...)
}
\arguments{
\item{x}{input character, json, or geojson}

\item{file}{file path to write to}

\item{...}{Further args}
}
\description{
Write inputs to a geojson file
}
\examples{
gj_write(x = '{"type": "Point", "coordinates": [-100, 80]}',
  (file <- tempfile()))
jsonlite::fromJSON(file)
}
