% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonlint-package.r
\docType{data}
\name{states}
\alias{states}
\title{This is the same data set from the ggplot2 library}
\format{A data.frame with 6 components, including "long", "lat", "group", "order",
"region", and "subregion" columns specifying polygons for each US state.}
\description{
This is the same data set from the ggplot2 library
}
\keyword{data}

