% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonlint.R
\name{geojson_lint}
\alias{geojson_lint}
\title{Validate GeoJSON using geojson_lint.com}
\usage{
geojson_lint(x, ...)
}
\arguments{
\item{x}{Input, a geojson character string, json object, or file or
url pointing to one of the former}

\item{...}{curl options passed on to \code{\link[httr]{GET}} or
\code{\link[httr]{POST}}}
}
\description{
Validate GeoJSON using geojson_lint.com
}
\details{
Uses the web service at \url{http://geojsonlint.com}
}
\examples{
\dontrun{
# From a json character string
geojson_lint(x = '{"type": "Point", "coordinates": [-100, 80]}') # good
geojson_lint(x = '{"type": "Rhombus", "coordinates": [[1, 2], [3, 4], [5, 6]]}') # bad

# A file
file <- system.file("examples", "zillow_or.geojson", package = "geojsonio")
geojson_lint(x = as.location(file))

# A URL
url <- "https://raw.githubusercontent.com/glynnbird/usstatesgeojson/master/california.geojson"
geojson_lint(as.location(url))
}
}

